move SettingsForm component into its own folder
