+++ src/pages/StartPage.tsx
@@ -3,8 +3,8 @@ import { useNavigate } from 'react-router-dom';
 import Header from '../components/Header/Header';
 import Button from '../components/Button/Button';
 import Card from '../components/Card/Card';
-import SettingsForm from '../components/SettingsForm';
-import Modal from '../components/Modal';
+import SettingsForm from '../components/SettingsForm/SettingsForm';
+import Modal from '../components/Modal/Modal';
 import { useQuizData } from '../hooks/useQuizData';
 import { useGameSettings } from '../context/GameSettingsContext';
 import type { GameSettings } from '../types';
